Factor out FileSink
